Handle that an operator numeric can be null
Future APNs will not use mcc+mnc and mvno data but instead be
set with only carrier id. This means that the definition of the
sim cards are solely in the carrier list and only have to be
changed in one place when there are new additions or change
of ownership etc.

When only carrier id is used to define an APN,
ApnSetting#getOperatorNumeric() will return null. DcTracker need
to handle this case and instead look for carrier id when needed.

Test: Add an APN with only carrier id and no mcc/mnc data
Bug: 217871446
Change-Id: I6e3a82fcd1a80710b6aa6fa509f0957244d7a927
Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>